Assembler: Fix default section not being reset to .text after an include
If you .included C code with a static variable, the end of the compiled ASM would leave off with the .section set to .sbss. If there was any more ASM code after the included C, then it would be ignored by the assembler as it was placed in the BSS section.
Now, the assembler resets the section to .text after every include so this doesn't need to be handled done manually.