Fix issue when inserting custom C code into a REL via a patch
The issue is that the C code uses Ghidra-style symbols, but once converted into ASM for the REL, the relocation needs to use framework.map-style symbols in order for the patcher to be able to find the relocated symbol's address.
To fix this, I had the assembler convert from Ghidra-style to framework.map-style symbol names when creating the diff.