Fix bug in switch_op class
If you set the SwOp's continuous param to false, and also specified an event to run when activated, the SwOp would start the event but never properly finish it.
This reorganizes the code to decouple the event checking from the switch checking so they can each work independently from one another. The code is also more readable now.