Completely refactor switch_op
This fixes a number of bugs in the class and makes it function much more similarly to the vanilla andsw2 class by using states.

For example, commit 430c8adf4a1ed71b0406ddad11b44bd726dbbc40 introduced a bug where the SwOp would fail to start its event if a different event was ongoing when its switch was set. That's fixed.

Commit 0a573e5029eace69d0b01cf83deed17b227bdc31 had a different bug where the SwOp would replay the event every time you exited and re-entered the room. That's now also fixed.